+++ packages/xbell/src/types/page.ts
@@ -10,6 +10,7 @@ import type {
   Download,
   TimeoutOptions,
 } from './pw';
+import { Page as PWPage } from 'playwright-core';
 import type { Locator } from './locator';
 import type { FrameLocator } from './frame-locator';
 import type { ElementHandle } from './element-handle';
@@ -68,6 +69,10 @@ export interface Page<BrowserExtensionArg = {}> extends CommonPage {
     waitUntil?: 'load' | 'domcontentloaded' | 'networkidle' | 'commit';
   }): Promise<Response | null>;
   waitForFileChooser(): Promise<FileChooser>;
+  $: PWPage['$'];
+  on: PWPage['on'];
+  off: PWPage['off'];
+  waitForEvent: PWPage['waitForEvent'];
 }
 
 export type PageMethods = Omit<Page, 'mouse' | 'keyboard'>;