+++ packages/color/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xbell/color
 
+## 0.8.9
+
+### Patch Changes
+
+- 0364253: Browser toMatchImageSnapshot
+
+## 0.8.8
+
+### Patch Changes
+
+- 0f95c96: Some apis
+
 ## 0.8.7
 
 ### Patch Changes