+++ packages/vue2/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @xbell/vue2
 
+## 0.8.9
+
+### Patch Changes
+
+- 0364253: Browser toMatchImageSnapshot
+- Updated dependencies [0364253]
+  - @xbell/snapshot@0.8.9
+  - @xbell/assert@0.8.9
+  - xbell@0.8.9
+
+## 0.8.8
+
+### Patch Changes
+
+- 0f95c96: Some apis
+- Updated dependencies [0f95c96]
+  - @xbell/snapshot@0.8.8
+  - @xbell/assert@0.8.8
+  - xbell@0.8.8
+
 ## 0.8.7
 
 ### Patch Changes