+++ packages/xbell/CHANGELOG.md
@@ -1,5 +1,31 @@
 # xbell
 
+## 0.8.9
+
+### Patch Changes
+
+- 0364253: Browser toMatchImageSnapshot
+- Updated dependencies [0364253]
+  - @xbell/code-stack@0.8.9
+  - @xbell/coverage@0.8.9
+  - @xbell/reporter@0.8.9
+  - @xbell/snapshot@0.8.9
+  - @xbell/assert@0.8.9
+  - @xbell/color@0.8.9
+
+## 0.8.8
+
+### Patch Changes
+
+- 0f95c96: Some apis
+- Updated dependencies [0f95c96]
+  - @xbell/code-stack@0.8.8
+  - @xbell/coverage@0.8.8
+  - @xbell/reporter@0.8.8
+  - @xbell/snapshot@0.8.8
+  - @xbell/assert@0.8.8
+  - @xbell/color@0.8.8
+
 ## 0.8.7
 
 ### Patch Changes